web_server/requirements.txt
certifi==2020.12.5
click==7.1.2
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
numpy==1.19.4
Pillow==8.0.1
Werkzeug==1.0.1
Keras==2.4.3
Keras-Preprocessing==1.1.2
numpy==1.19.4
opencv-contrib-python==4.5.1.48
scikit-image==0.17.2
scikit-learn==0.24.1
scipy==1.5.3
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
h5py==2.10.0
grpcio==1.32.0
opencv-python-headless==4.5.1.48
